+++ examples/ggeffects-test.R
@@ -1,6 +1,8 @@
 # test ggeffects
 
 library(ggeffects)
+# tested using latgest version, 1.2.2.12
+
 library(nestedLogit)
 
 data("Womenlf", package = "carData")
@@ -13,3 +15,13 @@ m <- nestedLogit(partic ~ hincome + children,
 )
 
 ggpredict(m, c("hincome", "children")) |> plot()
+
+# Data were 'prettified'. Consider using `terms="hincome [all]"` to get smooth plots.
+# Error: At least one term specified in `terms` is no valid model term.
+
+ggeffect(m, terms = "hincome[5:40, by=5]")
+
+# Can't compute marginal effects, 'effects::Effect()' returned an error.
+#
+# Reason: object 'prior.weights' not found
+# You may try 'ggpredict()' or 'ggemmeans()'.