+++ NEWS.md
@@ -1,6 +1,7 @@
 # nestedLogit 0.3.1
 
 * correct buglet in `as.data.frame` method when `newdata` has one column
+* `predict.nestedLogit()` now includes the `newdata` data frame as an additional component (`.data`) in its result. Consequently, the `newdata` argument is no longer required in the `as.data.frame` method.
 
 # nestedLogit 0.3.0